+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -11,7 +11,7 @@ create table products
 (   id bigserial    primary key,
     title           varchar(255),
     category_id     bigint references categories (id),
-    price           int,
+    price           numeric,
     height          int,
     weight          int,
     description     varchar(255),
@@ -32,7 +32,7 @@ create table orders
 (
     id              bigserial primary key,
     username        varchar(255) not null,
-    total_price     int not null,
+    total_price     numeric not null,
     address         varchar(255),
     phone           varchar(255),
     created_at      timestamp default current_timestamp,
@@ -45,8 +45,8 @@ create table  order_items
     product_id          bigint not null references products (id),
     order_id            bigint not null references orders (id),
     quantity            int not null,
-    price_per_product   int not null,
-    price               int not null,
+    price_per_product   numeric not null,
+    price               numeric not null,
     created_at          timestamp default current_timestamp,
     updated_at          timestamp default current_timestamp
 );
